Move .generate tests to separate file.
